Move DrawTargetExt and adapters from core to e-g
