+++ core/src/prelude.rs
@@ -2,11 +2,10 @@
 
 #[doc(no_inline)]
 pub use crate::{
-    draw_target::{DrawTarget, DrawTargetExt},
+    draw_target::DrawTarget,
     drawable::{Drawable, Pixel},
     geometry::{Dimensions, OriginDimensions, Point, Size},
     image::ImageDrawable,
-    iterator::{ContiguousIteratorExt, IntoPixels, PixelIteratorExt},
     pixelcolor::{
         raw::{RawData, ToBytes as _},
         GrayColor, IntoStorage, PixelColor, RgbColor, WebColors,
