+++ src/draw_target/cropped.rs
@@ -1,37 +1,41 @@
 use crate::{
-    draw_target::DrawTarget,
-    geometry::{Dimensions, Point},
-    iterator::PixelIteratorExt,
+    draw_target::{DrawTarget, DrawTargetExt, Translated},
+    geometry::{Dimensions, Point, Size},
     primitives::Rectangle,
     Pixel,
 };
 
-/// Translated draw target.
+/// Cropped draw target.
 ///
-/// Created by calling [`translated`] on any [`DrawTarget`].
-/// See the [`translated`] method documentation for more.
+/// Created by calling [`cropped`] on any [`DrawTarget`].
+/// See the [`cropped`] method documentation for more.
 ///
 /// [`DrawTarget`]: trait.DrawTarget.html
-/// [`translated`]: trait.DrawTargetExt.html#tymethod.translated
+/// [`cropped`]: trait.DrawTargetExt.html#tymethod.cropped
 #[derive(Debug)]
-pub struct Translated<'a, T>
+pub struct Cropped<'a, T>
 where
     T: DrawTarget,
 {
-    parent: &'a mut T,
-    offset: Point,
+    parent: Translated<'a, T>,
+    size: Size,
 }
 
-impl<'a, T> Translated<'a, T>
+impl<'a, T> Cropped<'a, T>
 where
     T: DrawTarget,
 {
-    pub(super) fn new(parent: &'a mut T, offset: Point) -> Self {
-        Self { parent, offset }
+    pub(super) fn new(parent: &'a mut T, area: &Rectangle) -> Self {
+        let area = area.intersection(&parent.bounding_box());
+
+        Self {
+            parent: parent.translated(area.top_left),
+            size: area.size,
+        }
     }
 }
 
-impl<T> DrawTarget for Translated<'_, T>
+impl<T> DrawTarget for Cropped<'_, T>
 where
     T: DrawTarget,
 {
@@ -42,62 +46,55 @@ where
     where
         I: IntoIterator<Item = Pixel<Self::Color>>,
     {
-        self.parent
-            .draw_iter(pixels.into_iter().translate(self.offset))
+        self.parent.draw_iter(pixels)
     }
 
     fn fill_contiguous<I>(&mut self, area: &Rectangle, colors: I) -> Result<(), Self::Error>
     where
         I: IntoIterator<Item = Self::Color>,
     {
-        let area = area.translated(self.offset);
-        self.parent.fill_contiguous(&area, colors)
+        self.parent.fill_contiguous(area, colors)
     }
 
     fn fill_solid(&mut self, area: &Rectangle, color: Self::Color) -> Result<(), Self::Error> {
-        let area = area.translated(self.offset);
-        self.parent.fill_solid(&area, color)
-    }
-
-    fn clear(&mut self, color: Self::Color) -> Result<(), Self::Error> {
-        self.parent.clear(color)
+        self.parent.fill_solid(area, color)
     }
 }
 
-impl<T> Dimensions for Translated<'_, T>
+impl<T> Dimensions for Cropped<'_, T>
 where
     T: DrawTarget,
 {
     fn bounding_box(&self) -> Rectangle {
-        self.parent.bounding_box().translated(-self.offset)
+        Rectangle::new(Point::zero(), self.size)
     }
 }
 
 #[cfg(test)]
 mod tests {
-    // NOTE: `crate` cannot be used here due to circular dependency resolution behavior.
-    use embedded_graphics::{
+    use crate::{
         draw_target::{DrawTarget, DrawTargetExt},
         geometry::Dimensions,
         geometry::{Point, Size},
         mock_display::MockDisplay,
         pixelcolor::BinaryColor,
-        primitives::Rectangle,
-        transform::Transform,
-        Pixel,
+        primitives::{Primitive, Rectangle},
+        style::PrimitiveStyle,
+        Drawable, Pixel,
     };
 
     #[test]
     fn draw_iter() {
         let mut display = MockDisplay::new();
 
-        let mut translated = display.translated(Point::new(2, 3));
+        let area = Rectangle::new(Point::new(2, 3), Size::new(10, 10));
+        let mut cropped = display.cropped(&area);
 
         let pixels = [
             Pixel(Point::new(0, 0), BinaryColor::On),
             Pixel(Point::new(1, 2), BinaryColor::Off),
         ];
-        translated.draw_iter(pixels.iter().copied()).unwrap();
+        cropped.draw_iter(pixels.iter().copied()).unwrap();
 
         display.assert_pattern(&[
             "    ", //
@@ -113,7 +110,8 @@ mod tests {
     fn fill_contiguous() {
         let mut display = MockDisplay::new();
 
-        let mut translated = display.translated(Point::new(3, 2));
+        let area = Rectangle::new(Point::new(3, 2), Size::new(10, 10));
+        let mut cropped = display.cropped(&area);
 
         let colors = [
             1, 1, 1, 1, 1, //
@@ -122,7 +120,7 @@ mod tests {
             1, 0, 1, 0, 1, //
         ];
         let area = Rectangle::new(Point::new(1, 2), Size::new(5, 4));
-        translated
+        cropped
             .fill_contiguous(&area, colors.iter().map(|c| BinaryColor::from(*c != 0)))
             .unwrap();
 
@@ -142,10 +140,11 @@ mod tests {
     fn fill_solid() {
         let mut display = MockDisplay::new();
 
-        let mut translated = display.translated(Point::new(1, 3));
+        let area = Rectangle::new(Point::new(1, 3), Size::new(10, 10));
+        let mut cropped = display.cropped(&area);
 
         let area = Rectangle::new(Point::new(2, 1), Size::new(3, 4));
-        translated.fill_solid(&area, BinaryColor::On).unwrap();
+        cropped.fill_solid(&area, BinaryColor::On).unwrap();
 
         display.assert_pattern(&[
             "      ", //
@@ -162,11 +161,15 @@ mod tests {
     #[test]
     fn clear() {
         let mut display = MockDisplay::new();
-        let mut translated = display.translated(Point::new(1, 3));
-        translated.clear(BinaryColor::On).unwrap();
+
+        let area = Rectangle::new(Point::new(1, 3), Size::new(3, 4));
+        let mut cropped = display.cropped(&area);
+        cropped.clear(BinaryColor::On).unwrap();
 
         let mut expected = MockDisplay::new();
-        expected.clear(BinaryColor::On).unwrap();
+        area.into_styled(PrimitiveStyle::with_fill(BinaryColor::On))
+            .draw(&mut expected)
+            .unwrap();
 
         display.assert_eq(&expected);
     }
@@ -174,13 +177,29 @@ mod tests {
     #[test]
     fn bounding_box() {
         let mut display: MockDisplay<BinaryColor> = MockDisplay::new();
+
+        let size = Size::new(3, 4);
+        let area = Rectangle::new(Point::new(1, 3), size);
+        let cropped = display.cropped(&area);
+
+        assert_eq!(cropped.bounding_box(), Rectangle::new(Point::zero(), size));
+    }
+
+    #[test]
+    fn bounding_box_is_clipped() {
+        let mut display: MockDisplay<BinaryColor> = MockDisplay::new();
         let display_bb = display.bounding_box();
 
-        let translated = display.translated(Point::new(1, 3));
+        let top_left = Point::new(10, 20);
+        let size = Size::new(1000, 1000);
+        let area = Rectangle::new(top_left, size);
+        let cropped = display.cropped(&area);
+
+        let expected_size = display_bb.size - Size::new(top_left.x as u32, top_left.y as u32);
 
         assert_eq!(
-            display_bb.translate(-Point::new(1, 3)),
-            translated.bounding_box()
+            cropped.bounding_box(),
+            Rectangle::new(Point::zero(), expected_size),
         );
     }
 }
