+++ src/iterator/contiguous.rs
